Move ALE benchmark helpers into adapter
